Refactor Entra interfaces to common project namespace
Move Entra (Microsoft Graph) integration interfaces and credentials to Wayd.Common.Application.Interfaces.ExternalPeople. Update all references and usages accordingly. Rename methods and variables in MicrosoftGraphService for clarity. Apply minor code style improvements. Remove obsolete project reference to Wayd.Integrations.MicrosoftGraph.
