+++ Wayd.Common/src/Wayd.Common.Application/Interfaces/ExternalPeople/EntraConnectionCredentials.cs
@@ -1,4 +1,4 @@
-namespace Wayd.Integrations.MicrosoftGraph;
+namespace Wayd.Common.Application.Interfaces.ExternalPeople;
 
 /// <summary>
 /// Per-connection credentials used to build a transient <see cref="Microsoft.Graph.GraphServiceClient"/>
