+++ Wayd.Services/Wayd.AppIntegration/src/Wayd.AppIntegration.Application/Wayd.AppIntegration.Application.csproj
@@ -11,7 +11,6 @@
         <ProjectReference Include="..\..\..\..\Wayd.Common\src\Wayd.Common.Application\Wayd.Common.Application.csproj" />
         <ProjectReference Include="..\..\..\..\Wayd.Common\src\Wayd.Common\Wayd.Common.csproj" />
         <ProjectReference Include="..\..\..\..\Wayd.Integrations\src\Wayd.Integrations.Abstractions\Wayd.Integrations.Abstractions.csproj" />
-        <ProjectReference Include="..\..\..\..\Wayd.Integrations\src\Wayd.Integrations.MicrosoftGraph\Wayd.Integrations.MicrosoftGraph.csproj" />
         <ProjectReference Include="..\Wayd.AppIntegration.Domain\Wayd.AppIntegration.Domain.csproj" />
     </ItemGroup>
